renamed `new_intersect_merge.py` and `merge_consensus_jobs.py`
