+++ cvat/apps/consensus/consensus_reports.py
@@ -26,7 +26,7 @@ from cvat.apps.consensus.models import (
     ConsensusReport,
     ConsensusSettings,
 )
-from cvat.apps.consensus.new_intersect_merge import IntersectMerge
+from cvat.apps.consensus.intersect_merge import IntersectMerge
 from cvat.apps.dataset_manager.util import bulk_create
 from cvat.apps.engine import serializers as engine_serializers
 from cvat.apps.engine.models import Job, Task, User
